[Chore] Fix 32 lint issues in test files (#232)
Fix all lint findings across 19 test files:
- gofmt: fix formatting in 13 files
- goconst: extract repeated string literals to constants
  (testMessage, testFederationID, testLocalMember, testHealthPath)
  and reuse existing testAddrEWMA constant in maglev/ringhash tests
- gosec G306: change file permissions from 0644 to 0600 in tls_test
- gosec G115: eliminate int->int32 overflow by using int32 loop vars
- staticcheck SA5011: use t.Fatal instead of t.Error before deref
- unparam: remove unused caKey return from generateTestCertificates,
  remove unused goroutine parameter in sticky concurrent test
- ineffassign: use blank identifier for unused return value
- revive: rename shadowed built-in 'copy' to 'cloned'

Resolves #232

+++ internal/agent/router/format_test.go
@@ -49,11 +49,11 @@ func TestFormatEndpointKey(t *testing.T) {
 func TestFormatEndpointKey_Concurrent(t *testing.T) {
 	done := make(chan bool)
 
-	for i := 0; i < 100; i++ {
-		go func(id int) {
-			for j := 0; j < 100; j++ {
+	for i := int32(0); i < 100; i++ {
+		go func(id int32) {
+			for j := int32(0); j < 100; j++ {
 				address := "10.0.0.1"
-				port := int32(id*100 + j)
+				port := id*100 + j
 				result := formatEndpointKey(address, port)
 				if result == "" {
 					t.Error("formatEndpointKey returned empty string")
